% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_path}
\alias{add_path}
\title{Funkcja zwraca zbiorcza ramke danych ~paths uaktualniona o sciezke z ~newPath}
\usage{
add_path(newPath, paths)
}
\arguments{
\item{newPath}{ramka danych z kolumnami $from, $to, $hour opisujaca poruszanie sie
konkretnego zwiedzajacego po galerii}

\item{paths}{ramka danych z informacjami o wszystkich przejsciach w galerii}
}
\description{
Funkcja zwraca zbiorcza ramke danych ~paths uaktualniona o sciezke z ~newPath
}

