% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataFiles.R
\name{create_dataFiles}
\alias{create_dataFiles}
\title{Funkcja przetwarza pliki z informacjami o zalogowaniach danego dnia na ramki danych ze
zliczonymi sciezkami i wynikowe pliki zapisuje w podanym folderze}
\usage{
create_dataFiles(allExhibits, rawDataFolder, returnFolder)
}
\arguments{
\item{allExhibits}{wektor character numerow wszystkich urzadzen}

\item{rawDataFolder}{sciezka dostepu do folderu zawierajacego pliki do obrobki}

\item{returnFolder}{sciezka dostepu do folderu, w ktorym maja byc zapisane pliki wynikowe}
}
\description{
Funkcja przetwarza pliki z informacjami o zalogowaniach danego dnia na ramki danych ze
zliczonymi sciezkami i wynikowe pliki zapisuje w podanym folderze
}

