% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_emptyPaths.R
\name{create_emptyPaths}
\alias{create_emptyPaths}
\title{Funkcja tworzy ramke danych ze wszystkimi mozliwymi przejsciami miedzy eksponatami}
\usage{
create_emptyPaths(allExhibits)
}
\arguments{
\item{allExhibits}{wektor character numerow eksponatow}
}
\description{
Funkcja tworzy ramke danych ze wszystkimi mozliwymi przejsciami miedzy eksponatami
}

