% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane_all.R
\docType{data}
\name{dane_all}
\alias{dane_all}
\title{Zbiór scieżek dla każdego dnia i dla poszczególnych godzin}
\usage{
data(dane_all)
}
\description{
Zbiór scieżek dla każdego dnia i dla poszczególnych godzin
}
\keyword{datasets}

