% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doStatystyk_funkcja.R
\name{doStatystyk_funkcja}
\alias{doStatystyk_funkcja}
\title{Tworzenie danych do statystyk}
\usage{
doStatystyk_funkcja(sciezka, out)
}
\arguments{
\item{sciezka}{ścieżka dostępu do katalogu ze ścieżkami}

\item{out}{ścieżka do katalogu, w którym mają zostać zapisane dane (jako plik
doStatystyk.Rda)}
}
\description{
Funkcja służy do tworzenia danych z długościami i czasami ścieżek dla każdego dnia
}

