% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClosest.R
\name{findClosest}
\alias{findClosest}
\title{Eksponat najbliższy aktualnej pozycji kursora}
\usage{
findClosest(coords, slownik)
}
\arguments{
\item{coords}{Współrzędne kursora}

\item{slownik}{Słownik urządzeń}
}
\value{
Funkcja zwraca dwuelementowy wektor typu string zawierający informacje o nazwie i numerze eksponatu.
}
\description{
Funkcja zwraca informacje o eksponacie znajdującym się najbliżej kursora.
}

