% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., cols = 1, titlesize = 20,
  title = "Analiza pojedynczych ścieżek w wybranych dniach")
}
\arguments{
\item{...}{wykresy do złączenia}

\item{cols}{liczba kolumn z wykresami}

\item{titlesize}{wielkość czcionki tytułu}

\item{title}{tytuł wykresu}
}
\value{
Wykresy złączone według specyfikacji.
}
\description{
Funkcja służy do łączenia wykresów w jeden wykres.
}

