% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_etykiety_nr.R
\name{plot_etykiety_nr}
\alias{plot_etykiety_nr}
\title{Funkcja służy do rysowania etykiet z numerami urządzeń}
\usage{
plot_etykiety_nr(dane_urz, przes = c(0, 0), cex = 0.7, col = "white",
  rozmiary = c(1018, 886))
}
\arguments{
\item{dane_urz}{ramka danych z kolumnami x i y odpowiadającymi współrzędnym urządzeń oraz kolumną nr - numerem urządzenia}

\item{przes}{2 - el. wektor przesunięcia etykiet, dodatnie wartości oznaczają odpowiednio: w lewo, w dół, ujemne prawo, góra
domyślnie etykiety są ustawione na środku linii łączącej punkty}

\item{cex}{wielkość tekstu etykiet}

\item{col}{kolor tekstu etykiet}

\item{rozmiary}{rozmiary planu, na który nanosimy punkty}
}
\description{
Funkcja wymaga uprzedniego wywołania plot.new (np. przez wywołanie funkcji plot_mapa)
}

