% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_liczby.R
\name{plot_liczby}
\alias{plot_liczby}
\title{Funkcja służy do rysowania etykiet z liczbami nad liniami łączącymi urządzenia}
\usage{
plot_liczby(dane_pocz, dane_kon, wartosci, przes = c(-2, -5), cex = 0.9,
  col = "dodgerblue4", rozmiary = c(1018, 886))
}
\arguments{
\item{dane_pocz}{ramka danych z kolumnami x i y odpowiadającymi współrzędnym początkowych urządzeń}

\item{dane_kon}{ramka danych z kolumnami x i y odpowiadającymi współrzędnym końcowych urządzeń}

\item{wartosci}{wektor o długości równej długości powyższych ramek danych z etykietami wyświetlanymi nad liniami}

\item{przes}{2 - el. wektor przesunięcia etykiet, dodatnie wartości oznaczają odpowiednio: w lewo, w dół, ujemne prawo, góra
domyślnie etykiety są ustawione na środku linii łączącej punkty}

\item{cex}{wielkość tekstu etykiet}

\item{col}{kolor tekstu etykiet}

\item{rozmiary}{rozmiary planu, na który nanosimy punkty}
}
\description{
Funkcja wymaga uprzedniego wywołania plot.new (np. przez wywołanie funkcji plot_mapa)
}

