% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paths4.R
\name{plot_paths2}
\alias{plot_paths2}
\title{Funkcja służy do rysowania scieżek z danych}
\usage{
plot_paths2(data, slownik, col = "#160773", lwd = 3.5, rozmiary = c(1018,
  886))
}
\arguments{
\item{data}{ramka danych z kolumnami tekstowymi from i to określającymi urządzenia początkowe i końcowe oraz numeryczną
kolumnętotal}

\item{slownik}{ramka danych z kolumnami x i y określającymi współrzędne urządzenia i  kolumnę nr określającą 
numer urządzenia}

\item{col}{kolor linii}

\item{lwd}{szerekość linii}

\item{rozmiary}{rozmiary mapy}
}
\description{
Funkcja służy do rysowania scieżek z danych
}

