% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paths_graph.R
\name{plot_paths_graph}
\alias{plot_paths_graph}
\title{Funkcja służy do rysowania scieżek z danych}
\usage{
plot_paths_graph(data, slownik)
}
\arguments{
\item{data}{ramka danych z kolumnami tekstowymi from i to określającymi urządzenia początkowe i końcowe oraz numeryczną 
kolumną total}

\item{slownik}{ramka danych z kolumnami x i y określającymi współrzędne urządzeń i kolumną nr określającą numer urządzenia}
}
\description{
Funkcja służy do rysowania scieżek z danych
}

