% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polaczenia.R
\name{plot_polaczenia}
\alias{plot_polaczenia}
\title{Funkcja służy do rysowania polaczen}
\usage{
plot_polaczenia(dane_pocz, dane_kon, wartosci = 0, col = "black",
  alpha = 0.2, lwd = 2, rozmiary = c(1018, 886))
}
\description{
Funkcja służy do rysowania polaczen
}

