% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_urzadz.R
\name{plot_urzadz}
\alias{plot_urzadz}
\title{Funkcja służy do rysowania urzadzen}
\usage{
plot_urzadz(dane_urz, pch = 19, cex = 0.5, col = "black",
  rozmiary = c(1018, 886))
}
\description{
Funkcja służy do rysowania urzadzen
}

