% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roznica_czasu.R
\name{roznica_czasu}
\alias{roznica_czasu}
\title{Różnica między czasami}
\usage{
roznica_czasu(czas_potem, czas_przedtem, units = "secs",
  format = "\%H:\%M:\%S")
}
\arguments{
\item{czas_potem}{wektor czasów późniejszych}

\item{czas_przedtem}{wektor czasów wcześniejszych}

\item{units}{jednostki, w jakich ma być zwracana różnica czasów}

\item{format}{format w jakim podane są czasy}
}
\value{
wektor z różnicami czasów w jednostkach określonych parametrem units.
}
\description{
Funkcja wyliczająca róznicę między czasami.
}

