% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysujBoxploty.R
\name{rysujBoxploty}
\alias{rysujBoxploty}
\title{Rysowanie boxplotów}
\usage{
rysujBoxploty(dni, summary_dl = c(min = 1, low = 3, med = 7, mean = 8.304,
  upper = 12, max = 53), summary_czas = c(min = 0, low = 578, med = 1582, mean
  = 2353, upper = 3076, max = 36220))
}
\arguments{
\item{dni}{wektor dni, których rozkład ma być wyrysowany}

\item{summary_dl}{wektor z charakterystykami rozkładu dla długości scieżek
dla okresu, z którym będziemy porównujemy (min, Ikwartyl, mediana, średnia, IIIkwartyl, max)}

\item{summary_czas}{wektor z charakterystykami rozkładu dla czasu scieżek
dla okresu, z którym będziemy porównujemy (min, Ikwartyl, mediana, średnia, IIIkwartyl, max)}
}
\description{
Funkcja służy do rysowania 3-częściowego wykresu (boxplot z rozkładem 
długości pojedynczej ścieżki, boxplot z rozkładem czasu pojedynczej ścieżki
i pusty wykres z informacjami o średniej dla wybranych wielkości
}

