% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciezki.R
\name{sciezka}
\alias{sciezka}
\title{Wyznaczanie popularnych ścieżek.}
\usage{
sciezka(filtered_data, n, first)
}
\arguments{
\item{filtered_data}{Ramka danych powstała w wyniku użycia funkcji filter_data}

\item{n}{Długość ścieżki}

\item{first}{Numer eksponatu, od którego ma się zaczynać ścieżka.}
}
\value{
Funkcja zwraca ramkę danych z kolejnymi przejściami między stacjami w formacie pasującym do pozostałych danych w pakiecie.
}
\description{
Funkcja służy do wyznaczani najbardziej popularnej scieżki
}

