% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slownik_urz.R
\docType{data}
\name{slownik_urz}
\alias{slownik_urz}
\title{Słownik urządzeń}
\usage{
data(slownik_urz)
}
\description{
Zbiór nazw, kodów i współrzędnych na mapie dla odpowiednich stacji
}
\keyword{datasets}

