% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitor_path.R
\name{visitor_path}
\alias{visitor_path}
\title{Funkcja wybiera sciezki dla zwiedzajacego (usuwam powroty do tego samego eksponatu) i
zwraca ramke danych z informacjami o przejsciach miedzy eksponatami z ich godzina}
\usage{
visitor_path(data, visitorId)
}
\arguments{
\item{data}{ramka danych z informacjami o wszystkich zalogowaniach do eksponatow
danego dnia}

\item{visitorId}{identyfikator odwiedzajacego}
}
\description{
Funkcja wybiera sciezki dla zwiedzajacego (usuwam powroty do tego samego eksponatu) i
zwraca ramke danych z informacjami o przejsciach miedzy eksponatami z ich godzina
}

