% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybraneDni_filter_data.R
\name{wybraneDni_filter_data}
\alias{wybraneDni_filter_data}
\title{Filtrowanie danych}
\usage{
wybraneDni_filter_data(weekdays = c(1, 7), months = c(1, 12))
}
\arguments{
\item{weekdays}{wektor zawierający numeryczne wartości dni tygodnia,
które chcemy uwzględnić}

\item{months}{wektor zawierający numeryczne wartości miesięcy,
które chcemy uwzględnić}
}
\value{
zwraca wektor z datami typu dzień_miesiac (np. 21_02)
}
\description{
Funkcja służy do przetwarzania wybranych w opcjach aplikacji miesięcy i dni tygodnia do zbioru
dni rozumianych przez funkcję rysującą rysujBoxploty
}

