% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wybraneDni_filter_data_byday.R
\name{wybraneDni_filter_data_byday}
\alias{wybraneDni_filter_data_byday}
\title{Filtrowanie danych}
\usage{
wybraneDni_filter_data_byday(days)
}
\arguments{
\item{days}{wektor zawierający skrajne wartości zakresu dni, które chcemy uwzględnić}
}
\value{
zwraca wektor z datami typu dzień_miesiac (np. 21_02)
}
\description{
Funkcja służy do przetwarzania wybranych w opcjach aplikacji dni do zbioru
dni rozumianych przez funkcję rysującą rysujBoxploty
}

