% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Funkcja służy do filtrowania danych}
\usage{
filter_data(weekdays = c(1, 7), months = c(1, 12), hours = c(9, 20))
}
\arguments{
\item{weekdays}{Wektor zawierający numeryczne wartości dni tygodnia,
które chcemy uwzględnić.}

\item{months}{Wektor zawierający numeryczne wartości miesięcy,
które chcemy uwzględnić.}

\item{hours}{Wektor dwuelementowy zawierający informacje o przedziale godzin, który chcemy uwzględnić}
}
\description{
Funkcja służy do filtrowania danych
}

