% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data_byday.R
\name{filter_data_byday}
\alias{filter_data_byday}
\title{Funkcja służy do filtrowania danych po miesiącach, dniach tygodnia i godzinach.}
\usage{
filter_data_byday(days, hours = c(9, 20))
}
\arguments{
\item{days}{Wektor zawierający skrajne wartości zakresu dni, które chcemy uwzględnić}

\item{hours}{Wektor dwuelementowy zawierający informacje o przedziale godzin, który chcemy uwzględnić}
}
\description{
Funkcja służy do filtrowania danych po miesiącach, dniach tygodnia i godzinach.
}

