% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClosest.R
\name{findClosest}
\alias{findClosest}
\title{Funkcja zwraca informacje o stacji znajdującej się najbliżej kursora}
\usage{
findClosest(coords, slownik)
}
\arguments{
\item{coords}{Współrzędne kursora}

\item{slownik}{Słownik urządzeń}
}
\description{
Funkcja zwraca informacje o stacji znajdującej się najbliżej kursora
}

