% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_etykiety.R
\name{plot_etykiety}
\alias{plot_etykiety}
\title{Funkcja służy do rysowania etykiet}
\usage{
plot_etykiety(dane_urz, przes = c(0, 0), cex = 0.5, col = "black",
  rozmiary = c(1018, 886))
}
\description{
Funkcja służy do rysowania etykiet
}

