% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciezki.R
\name{sciezka}
\alias{sciezka}
\title{Funkcja służy do wyznaczani najbardziej popularnej scieżki}
\usage{
sciezka(filtered_data, n, first)
}
\arguments{
\item{filtered_data}{Ramka danych powstała w wyniku użycia funkcji filter_data}

\item{n}{Długość ścieżki}
}
\description{
Funkcja służy do wyznaczani najbardziej popularnej scieżki
}

